+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.39 (Pending):
+* Extract AbstractSharedTest from AbstractServerTest to allow reuse of the test
+  infrastructure in client-side (i.e. GWT) and non-JEE frameworks.
 * Move AbstractModule to the shared package but create a deprecated old AbstractModule
   that extends the new AbstractModule.